Assignment 3/requirements.txt
openai
groq
requests
google-generativeai
cohere
python-dotenv
streamlit